requirements.txt
asgiref==3.7.2
awslogs==0.14.0
boto3==1.18.18
botocore==1.21.18
channels==3.0.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
Django==3.0.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-jwt==0.3.3
djangorestframework==3.13.0
djongo==1.3.6
dnspython==2.1.0
et-xmlfile==1.1.0
fiscalyear==0.3.2
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
idna==3.2
jmespath==0.10.0
jsonfield==3.1.0
Markdown==3.3.4
numpy==1.24.4
openpyxl==3.0.9
pandas==1.3.5
PyJWT==2.3.0
pymongo==3.12.0
python-dateutil==2.8.2
pytz==2021.1
razorpay==1.2.0
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.2.4
textract-trp==0.1.3
urllib3==1.26.7
XlsxWriter==3.0.1
easygui==0.98.3
tk==0.1.0
opencv-python==4.6.0.66
Pillow==10.3.0
pytesseract==0.3.8
selenium==4.5.0
urllib3==1.26.7
webdriver-manager==3.4.2
watchtower==3.0.1
random-password-generator==2.2.0
djangorestframework-simplejwt==5.2.2
schedule==1.2.0
drf_yasg==1.21.7
xlrd==1.2.0
pychromepdf==1.1
weasyprint==61.2







